requirements.txt
python-hcl2
pathlib